Swap folder for User and Read.me
Foi alterado de local a classe User para a pasta correta e o arquivo read.me foi reformulado para apresentação do trabalho.
