+++ ContactManagement.Api/Configurations/DependencyInjectionConfig.cs
@@ -10,7 +10,7 @@ public static class DependencyInjectionConfig
 {
     public static IServiceCollection ResolveDependencies(this IServiceCollection services, IConfiguration configuration)
     {
-        string connectionString = configuration.GetConnectionString("ConnectionRafael");
+        string connectionString = configuration.GetConnectionString("ConnectionLucas");
 
         services.AddTransient<IDbConnection>(db => new SqlConnection(connectionString));
         services.AddScoped<IContactServices, ContactServices>();
