+++ ContactManagement.Application/Interfaces/ITokenServices.cs
@@ -1,9 +1,8 @@
-using System.Threading.Tasks;
-using ContactManagement.Domain.Models;
+using ContactManagement.Domain.Entities;
 
 namespace ContactManagement.Application.Interfaces
 {
-    public interface ITokenService
+	public interface ITokenService
     {
         Task<string> GetToken(User user); // Método deve ser assíncrono
     }
