+++ README.md
@@ -1,35 +1,34 @@
-# Voce já deve ter o docker instalado e configurado na sua máquina
+<h1 align="center">Tech Challenge da Fase 1 - FIAP</h1>
 
-# Abra o terminal na pasta onde está o arquivo docker-compose.yml e execute **
+## :computer: Projeto
 
-docker-compose up --build
-Após finalizar de rodar, vá para o passo abaixo
+Uma API para cadastro de contatos regionais, considerando a persistência de dados e qualidade de software.
 
-# Acessar o banco de dados SQL com sua ferramenta de gerenciamento de Banco de dados, você precisa ter os acessos nescessário no BD para isto
-Se estiver utlizando o SQL Management Studio é só copiar o código abaixo e rodar, em outra plataforma como Dbeaver, será necessário rodar um item por vez
-Abra uma nova janela para rodar o script,  copie o código abaixo, da linha 11 a 35 e excute o código
--- 1 Cria o banco de dados 'Contacts'
-CREATE DATABASE ContactManagment;
-GO
+O principal dado a ser validado neste projeto é o DDD do telefone de cada contato, onde inclusive uma das rotas retorna somente contatos de uma determinada região de acordo com DDD.
 
--- 2 Seleciona o banco de dados 'ContactManagment'
-USE ContactManagment;
-GO
+Para utilização, é necessário uma autenticação de usuário que é validada por um Enumerador de permissões, que em diferentes níveis permitirá determinadas ações.
 
--- 3 Cria a tabela 'ContatctInformation'
-CREATE TABLE ContatctInformation (
-    Id INT IDENTITY(1,1) PRIMARY KEY,
-    FirstName NVARCHAR(70) NOT NULL,
-    LastName NVARCHAR(70) NOT NULL,
-    DDD CHAR(2) NOT NULL,
-    Phone INT NOT NULL,
-    Email NVARCHAR(100) NOT NULL
-);
-GO
+Houve um trabalho para validação de dados que são fornecidos para cadastro, onde é verificado se os campos são preenchidos e se são válidos de acordo com cada tipo de dado.
 
--- 4 Insere um contato na tabela 'ContatctInformation'
-INSERT INTO Contatcs (FirstName, LastName, DDD, Phone, Email)
-VALUES ('João', 'Silva', '11', 999999999, 'joao.silva@email.com');
-GO
--- 5 Lista os contatos inseridos
-SELECT TOP 1 * FROM ContatctInformation
+
+## :wrench: Tecnologias Utilizadas
+
+- **Documentação:** Swagger
+- **ORM:** Dapper (SqlServer)
+- **Autenticação:** JWTBearer e IdentityModel
+- **Testes unitários:** xUnit e Moq
+- **Logs:** ILog
+- **Validação de dados:** Fluent Validation
+
+## :arrow_forward: Como Utilizar
+
+Neste projeto é utilizado banco de dadosSql Server. 
+Após clonar o projeto, pode-se utilizar banco de dados local ou docker com a imagem e executar o script de banco de dados na raiz do projeto com o nome "scriptparacriarobanco.sql".
+
+docker: ```docker compose up --build```
+
+## Autores:
+- [Daniel Lopes](https://github.com/daniel-lopes-98435)
+- [Lucas Almeida](https://github.com/lugsdev)
+- [Marla Amoury](https://github.com/marlamoury)
+- [Rafael Sampaio](https://github.com/rafasampaio25)
